requirements.txt
Appium-Python-Client==3.1.1
pytest==7.4.3
pytest-html==4.1.1
selenium==4.15.2 